requirements.txt
requests>=2.32.0
pymongo>=4.8.0
pyspark>=3.5.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit>=1.36.0
python-dotenv>=1.0.1
prefect>=2.19.0
networkx>=3.3
scikit-learn>=1.5.0
statsmodels>=0.14.2
mlflow>=2.15.0
sentence-transformers>=3.0.1
prometheus-client>=0.20.0
